requirements.txt
Flask==2.3.0
flask-socketio==5.3.0
python-socketio==5.9.0
paho-mqtt==1.6.1
python-engineio==4.7.0
